rename to SqlStringUtil
